fix: BottomSheet no longer opens fully
Moved the contact async fetch in the Qualification Panel
to the FooterContent parent, this prevents `mui-bottom-sheet`
from miscalculating one of its style variables
