+++ react/Viewer/Footer/FooterContent.jsx
@@ -3,13 +3,19 @@ import PropTypes from 'prop-types'
 import { makeStyles } from '@material-ui/core/styles'
 
 import { isMobileApp } from 'cozy-device-helper'
+import { getReferencedBy, useQuery, models } from 'cozy-client'
 
+import { buildContactByIdsQuery } from '../queries'
 import { isValidForPanel } from '../helpers'
 import Sharing from './Sharing'
 import ForwardButton from './ForwardButton'
 import DownloadButton from './DownloadButton'
 import BottomSheetWrapper from './BottomSheetWrapper'
 import BottomSheetContent from './BottomSheetContent'
+
+const {
+  contact: { getDisplayName }
+} = models
 
 const useStyles = makeStyles(theme => ({
   footer: {
@@ -28,7 +34,23 @@ const FooterContent = ({ file, toolbarRef, disableSharing }) => {
   const FileActionButton = isMobileApp() ? ForwardButton : DownloadButton
   const actionButtonsRef = useRef()
 
-  if (isValidForPanel({ file })) {
+  const contactIds = getReferencedBy(file, 'io.cozy.contacts').map(
+    ref => ref.id
+  )
+  const contactByIdsQuery = buildContactByIdsQuery(contactIds)
+  const { data: contactList } = useQuery(contactByIdsQuery.definition, {
+    ...contactByIdsQuery.options,
+    enabled: contactIds.length > 0
+  })
+
+  const contactsFullname = Array.isArray(contactList)
+    ? contactList.map(contact => `${getDisplayName(contact)}`).join(', ')
+    : ''
+
+  if (
+    isValidForPanel({ file }) &&
+    (contactsFullname || contactIds.length === 0)
+  ) {
     return (
       <BottomSheetWrapper
         file={file}
@@ -38,6 +60,7 @@ const FooterContent = ({ file, toolbarRef, disableSharing }) => {
         <BottomSheetContent
           file={file}
           disableSharing={disableSharing}
+          contactsFullname={contactsFullname}
           ref={actionButtonsRef}
         />
       </BottomSheetWrapper>
